initial changes, incomplete
#64
#62
