+++ settings.py
@@ -6,11 +6,8 @@ from shutil import copyfile
 class Settings():
     
     def __init__(self, file: str):
-        """
-        initialization.
-        """
         self.settings : dict
-        self.settings_file = file
+        self._settings_file = file
 
         if not os.path.exists(self.settings_file):
             copyfile("defaultsettings.json", self.settings_file)
@@ -22,7 +19,7 @@ class Settings():
 
     def cleanup(self) -> None:
         """
-        Should be run when before exit. Outputs settings to file.
+        Outputs settings to file.
         """
-        with open (self.settings_file, "w") as f:
+        with open (self._settings_file, "w") as f:
             f.write(json.dumps(self.settings, indent=4))
